patch: sorry/fs-nvrpart
there is some extra noise in this diff due to the fact that i've got a different
dst scheme than you do at this point and i changed the tracing in devsd.
the imporant changes are in otherinit() and add9loadparts().  it looks to
be 20 lines of code for add9loadparts.

the cavet is that i have not tried this when using the disk nvr is on for a fs.
also, this points out that we should be able to specify partitions in config strings.
cPm0cache.fPm0fworm.

Reference: /n/sources/patch/sorry/fs-nvrpart
Date: Wed Mar 28 16:07:02 CES 2007
Signed-off-by: quanstro@quanstro.net