patch: applied/hushping
Ip/ping is chatty, even with the -q flag set. My
understanding of this flag is that it should stop all
output of ping from happening (that goes to stdout, as
opposed to stderr). In fact, ip/ping still outputs its
startup garble (sending bla bla) and, if it loses
packets, how many.

My expectation is that it should behave as e.g. grep -s
where no output happens to stdout, but the $status is
set.

This patch should fix this issue adequately (it's really
just 2 lines).

Reference: /n/sources/patch/applied/hushping
Date: Wed Apr  4 22:01:26 CES 2007