patch: sorry/termrc-superfluous-sysname-check
termrc checks both (~ $#sysname 0) and (~ $sysname '').
The latter would cover the former, but is not possible
for `{} to return a null string in any circumstances, so
the check is superfluous.

Notes:
the current termrc may be excessively cautious, but
it works and we've beaten this subject to death.

Reference: /n/sources/patch/sorry/termrc-superfluous-sysname-check
Date: Tue May 22 20:38:14 CES 2007
Signed-off-by: kmaglione@comcast.net
Reviewed-by: geoff