patch: saved/9pxeload-pxeorder
when pxebooting with multiple ethers available for pxebooting, after
the initial probe finds a .ini file the probe to find and partition the disks
does not skip ethers.  this means that if we found the .ini file on ether0
and there is more than one available ethernet ctlr, then ether1 will be the
first interface we try to boot from.  the simple fix is to only probe for the
first ether.

Notes:
Sun Sep 2 00:14:00 EDT 2007 geoff
    we've run into this problem before, but all the fixes
    have their own drawbacks.  we're about to revise 9load
    and will bear it in mind.

Reference: /n/sources/patch/saved/9pxeload-pxeorder
Date: Sun Jun 10 19:41:41 CES 2007
Signed-off-by: quanstro@quanstro.net
Reviewed-by: geoff