patch: saved/ftpfs-unix-spacefiles
Currently, ftpfs is broken with respect to files with spaces in their
filename.  We assume all files will be sent as a single string, while
this is not the case.  My fix is a kind of nasty hack, but I couldn't
think of a better way that didn't require a lot more restructuring
than I was willing to do.  Anyway, example of the problem:

ls /n/ftp/mp3/mp3
:
/n/ftp/mp3/mp3/Borsato
:

The expected output from ls is:
'/n/ftp/mp3/mp3/Marco Borsato'

With this patch, any filenames with spaces in them are fixed for
correct output:

'/n/ftp/mp3/mp3/thievery corporation - sounds from the thievery hi-fi/04. transcendence.mp3'

for example.

I reorganized the case handling as well, but that's more of a style
thing.  If you have other suggestions for changes that would make this
more suitable for incorporating as a patch, I'd love to hear them!

Notes:
Tue Oct 2 16:23:55 EDT 2007 geoff
    it is indeed a nasty hack and a moderately large diff for
    what one would hope to be a minor change.  it also seems that
    spaces are likely to cause problems elsewhere in ftp, client
    and server.  given that ftp is a complex, ugly, and old protocol,
    i'm inclined to document the limitation and call it a day.

    saved in case we decide to revisit this.

Reference: /n/sources/patch/saved/ftpfs-unix-spacefiles
Date: Tue Oct  2 22:12:59 CES 2007
Signed-off-by: devon.odell@gmail.com
Reviewed-by: geoff