patch: sorry/acme-mail-from
Let Acme mail pass "From:" headers through to the
message that gets sent to upas/marshal. I wasn't
clear at all on what the motivation was for treating
it as an additional "To:" line, but if that's
desirable, just replicate the TO case for FROM with
delit set to 0 instead of 1.

Notes:
Tue Nov 27 19:00:53 EST 2007 rsc
    The reason that acme treats From: as To: is so that
    if you are reading a message you can actually
    just edit the body and click Post to send a reply.
	that's why it's not correct to send the entire From: line.
	it's usually someone else's info!

    russ

Reference: /n/sources/patch/sorry/acme-mail-from
Date: Mon Oct 22 18:20:19 CES 2007
Signed-off-by: a@9srv.net
Reviewed-by: geoff