patch: applied/ape-errstr-mapping
fixed the errstr string for "directory not empty" - I assume the old string
"remove -- directory not empty" was for ken's FS. There are probably a dozen
of these to be done but this one was breaking cvs so worth doing.

the compare in the code does a strstr so the short form (for fossil)
should still work with ape, though I haven't tested this.

I guess I should try to collate the error strings between the
different file servers and try to keep ape in sync. one day maybe.

-Steve

Reference: /n/sources/patch/applied/ape-errstr-mapping
Date: Wed Nov  7 15:01:07 CET 2007
Signed-off-by: steve@quintile.net