patch: sorry/timerintr-wrong-proto
The function timerintr in this file had the wrong prototype.
I found this out whilst trying to build a cpu/auth server.
The format in the header file was
	void timerintr(Ureg *, uvlong)
The format in the actual source was
	void timerintr(Ureg *, Tval)
This is my first contrib to the kernel; sorry if it makes other stuff break.

Pietro Gagliardi

Notes:
Sat Mar 22 18:35:39 EDT 2008 geoff
    that's not the root problem.  replica hasn't been updating kernel
    sources correctly.  I've just touched
    n/sources/plan9/sys/src/9/port/portfns.h, so in half an hour, run
    replica/pull; you should get a new portfns.h with the correct
    prototype.

Reference: /n/sources/patch/sorry/timerintr-wrong-proto
Date: Sat Mar 22 20:19:23 CET 2008
Signed-off-by: pietro10@mac.com
Reviewed-by: geoff