patch: applied/ape-time-bsd-ext
Remove requirement to set _BSD_EXTENSION when including sys/time.h, it
is now part of the XSI standard and many apps require it.

Seems that the timezone part didn't make it into the standard and
seems to be deprecated, so I left it inside a _BSD_EXTENSION ifdef in
case any old code depends on it (I have't found any).

Reference: /n/sources/patch/applied/ape-time-bsd-ext
Date: Mon Jun 23 00:24:42 CES 2008
Signed-off-by: uriel99@gmail.com