patch: applied/httpd-schemes
This patch arose because I tried to run a wiki inside https instead of
http.

The key problem is that the "Location:" HTTP header must specify an
absolute URI, which means that whoever decides to do a redirect must
know not only the "pathname" and the hostname but also the "scheme"
(http or https) and the port number.  According to the current design
this information must be propagated not only throughout httpd.c but
also out to the helper programs via exec().

I've been running essentially this code for two years, and this patch
is based on merging in the intervening changes from sources,
rebuilding, and testing briefly.  However, I should state that there
are many features of httpd which I don't use, e.g., httpd.rewrite or
any of the magic programs such as the man-page generator (though I
have exercised wikipost).

Reference: /n/sources/patch/applied/httpd-schemes
Date: Mon Jun 23 05:21:00 CES 2008
Signed-off-by: davide+p9@cs.cmu.edu