patch: applied/webfsfixes
This patch fixes webfs not working on a nonstandard port for urls of the type:
http://www.url.com:8080:/asdfads
The count of subexpressions in the regular expression definitions in url.c was wrong.
It also fixes an error in the POST code which would post more than once for an authenticated
connection. I have tried this in the urls I care about, and the fixes seem not to break anything
else.
d

Reference: /n/sources/patch/applied/webfsfixes
Date: Mon Sep 29 11:40:41 CES 2008
Signed-off-by: paurea@gmail.com