patch: applied/9load-conf-error
Had a "warning: did not change ... in plan9.ini" message when using 9pxeload, but didnt on 9load. changeconf() copies the first bytes of the new configuration line to pref+1, but strchr() starts looking for the = at pref. But pref[0] is not initialized here. (It is later used to store an '\n'). This patch changes strchr(pref, '=') to strchr(pref+1, '='). This fixes the problem for me.

Notes:
Mon Oct 20 12:12:18 EDT 2008 geoff
    fixed differently.

Reference: /n/sources/patch/applied/9load-conf-error
Date: Sun Oct 19 16:12:56 CES 2008
Signed-off-by: cinap_lenrek@gmx.de
Reviewed-by: geoff