patch: applied/pops_tls
This is actually two related patches.  The patch to tlshand.c
fixes an inability to work with some POP servers.  In particular
some (e.g. the yahoo server that hosts AT&T service) do send
0 bytes with the certificate request, so I've removed the check
for 0.  The man page change just brings it up to date with the
source code, specifically the existance of the pops and apops
protocols.

Reference: /n/sources/patch/applied/pops_tls
Date: Wed Apr 15 18:13:59 CES 2009
Signed-off-by: blstuart@bellsouthnet