patch: applied/ups-update
UPS changed their tracking site again; updated /rc/bin/ups to
match. It now prints the attribution, the "on time message",
the tracking number (as UPS likes it), Type, Status, "Shipped
To" (which seems to be the city), Shipped/billed date, which
service was used, and the weight of the package. Or at least
that's what i get on the two tracking numbers i have to test.

There was also a comment about defeating caching proxies that
didn't seem to have an implementation, and a split of rfork's
e and n; i removed those.

Reference: /n/sources/patch/applied/ups-update
Date: Mon Jun 29 17:07:34 CES 2009
Signed-off-by: a@9srv.net