patch: applied/webls-root-fixed
Fixed ip/httpd/webls to use webroot as defined in httpd.c (i.e., with
httpd -w), instead of the hardcoded default.  This is also more in
line with manual page descriptions.  NOTE: this is also a fix to the
webls-root patch that I previously submitted -- it was missing changes
to file httpd.c; please refer to this instead

Reference: /n/sources/patch/applied/webls-root-fixed
Date: Mon Jul 27 02:23:56 CES 2009
Signed-off-by: akumar@mail.nanosouffle.net