patch: maybe/acme-label
Keeping track of many concurrent acme invocations in rio can be annoying.
While you can B2 'Local echo -n some_label > /dev/label' in a running acme,
there's no way to set the label on an acme window when starting it from the
command line or a script.

This patch adds a -L option that takes as its argument a label name for
the running acme. The label is valid only for the duration of the acme run,
preserving the parent rio window's label when acme terminates.

Reference: /n/sources/patch/maybe/acme-label
Date: Tue Oct 27 22:14:10 CET 2009
Signed-off-by: lyndon@orthanc.ca