patch: applied/httpd_support_javascript
http://mbgokhale.org/Media/slideshow.html?backgroundColor=rgb(0%2C%200%2C%200)&reflectionHeight=100&reflectionOffset=2&captionHeight=100&fullScreen=0&transitionIndex=2
Note that this looks like an invocation of a cgi-bin, but it is not -- and Plan 9 httpd will improperly refuse to pass the .html file back, failing with an HNoSearch error.
In fact, it should pass the file back, because the browser will apply the arguments to the javascripts. It appears to me that the error was
correct in the pre-javascript days but is now wrong.
The attached patch fixes the problem. It is a trivial deletion of two lines. I can now view javascript pages (such as those produced by Apple's iWeb tool)
with Plan 9 httpd.

Reference: /n/sources/patch/applied/httpd_support_javascript
Date: Fri Nov 27 03:45:34 CET 2009
Signed-off-by: rminnich@gmail.com