patch: applied/ape-dirread
post...
very strange bug, when doing a directory scan somtimes read(2) of the directory
returns -1 at EOF rather than zero. this confuses the ape readdir() code and so
we read off the end of the dir.

this only happens with my cifs(1) server but that uses lib9p and its dirgen()
function, which only does a respond(r, nil);.

As a result this is not really a fix but a workaround, however, I do think
its fair that the read system call could fail without causing a suicide in ape
so this change seems fair to me.

-Steve

Reference: /n/sources/patch/applied/ape-dirread
Date: Thu Jan 21 17:40:27 CET 2010
Signed-off-by: steve@quintile.net