patch: maybe/ndb-quotefmt
Use quotefmtinstall(); and %q when printing values in
ndb/query and ndb/ipquery.

Added quote parsing in libndb allowing strings quoted with " or ' and
embedded "" or '' inserts a single " or ' in the usual style; double quotes
don't feel like the plan9 way, but I kept support for backward compatability.

This useful for txt='' and hinfo='' strings.

-Steve

Reference: /n/sources/patch/maybe/ndb-quotefmt
Date: Fri Jan 22 13:35:32 CET 2010
Signed-off-by: steve@quintile.net