patch: applied/gs-mkfile
Added 'rm src/plan9.mak' to the clean and nuke clean section in the mkfile. Newly added drivers to the drivers section in the mkfile refuse to get compiled in, because an old plan9.mak will not be changed, although it should be. (Sorry for the poor gs-mkfile patch)
-- stevie

Reference: /n/sources/patch/applied/gs-mkfile
Date: Tue Feb  2 18:12:25 CET 2010
Signed-off-by: stevie@mordor.tip9ug.jp