patch: sorry/format-addfile
format(8) prints 'add file at clust n' for every file it adds to the fat partition. if this file. for example, is in a bind, it will print the name of real file, not the name that the file will have in the fat. make it print the name of the file in the fat (the same name passed in the addrname right below).
i had problems sending the other format-addfile patch, this is the correct one. sorry.

Notes:
Tue Jul 13 16:49:03 EDT 2010 geoff
    The current behaviour is fine.

Reference: /n/sources/patch/sorry/format-addfile
Date: Sat Mar 13 22:16:47 CET 2010
Signed-off-by: iru.muzgo@gmail.com
Reviewed-by: geoff