patch: applied/export-only-root
This patch fixes what turned out to be a longstanding problem.
The exportfs needs to export the root, not the namespace of the
server. This patch (from Russ) remounts '#s/boot' on /root and
then exports /root. This fix is needed for those users who
set up combined auth/cpu/fs systems, because the /mnt/factotum
on the auth server collides with the /mnt/factotum on cpu servers
that mount the file server (in the case that auth/cpu/fs are
the same).

Notes:
Fri Apr 2 12:47:50 EDT 2010 geoff
    I've applied the change to /n/sources/plan9/rc/bin/service/!tcp564.
    We don't actually run that script on any of our tcp 564 ports.

Reference: /n/sources/patch/applied/export-only-root
Date: Fri Apr  2 08:13:05 CES 2010
Signed-off-by: rminnich@gmail.com
Reviewed-by: geoff