patch: applied/kernel-validname-error
let validaddr send a note and raise an error so resources can be freed
on waserror() instead of calling pexit. the validaddr call in validname0
was removed, as the first byte should be already validated by the caller
(from syscalls). vmemchr will validate all following pages of the name
if it spans pages. let sysexec validnamedup() a save copy of the file
argument. the rest of the changes are simplifications of the error
handling mostly.

this patch replaces: kernel-sysexec-validnamedup, kernel-sysfile-validaddr,
kernel-segattach-validnamedup and kernel-namec-validaddr. removing the validaddr
calls for the string arguments doesnt work as the user can pass kernel
addresses!

Reference: /n/sources/patch/applied/kernel-validname-error
Date: Sun May  9 19:55:13 CES 2010
Signed-off-by: cinap_lenrek@gmx.de