patch: applied/sdata-cmd-0680
This adds support for the Silicon Image 0680/680A PATA133 controller,
as found on a Belkin F5U098 PCI card.  I haven't tested it with any
actual (storage) disks, but I was able to burn a CD which was readable
via a CD-ROM drive on a different controller, so I suspect disks will
work too.

Reference: /n/sources/patch/applied/sdata-cmd-0680
Date: Thu Jul 22 10:03:03 CES 2010
Signed-off-by: davide+p9@cs.cmu.edu