patch: maybe/cpu-tlspass
adds a new auth method "tlspass" that can be used with foreign
systems that have no other way than username/password to
authenticate users. server authentication is also provided in
this patch, but serves mostly as an example and should not be
used on plan9 cpu servers. (it will not work until an administrator
changes its listen script and passes a pem certificate on the
program arguments and has the rsa key in bootes factotum)

Reference: /n/sources/patch/maybe/cpu-tlspass
Date: Wed Oct  6 05:40:33 CES 2010
Signed-off-by: cinap_lenrek@gmx.de