patch: saved/remove-sprint-port
Removing all occurences of sprint() in the kernel source (port). This relates to a problem I had when compiling Ronald G. Minnich's 9vx repository on 64 bit Linux, which got me an error in some old version of sprint(). I decided, that it would be best, to remove this deprecated function in the main kernel source, on which the 9vx kernel source is directly based.

Further to this patch, I'm creating patches for all the folders in /sys/src/9.

Reference: /n/sources/patch/saved/remove-sprint-port
Date: Fri Feb 25 21:56:22 CET 2011
Signed-off-by: 20h@r-36.net