patch: bad/softcurs-pixel-dropping
software cursor was not removed from screen when loadimage
was used to fill on screen image content. this patch adds
hwdraw() calls to cloadmemimage and loadmemimage that are
used to remove the softcursor from the screen if the dst
image overlaps the position of the softcursor.

hwdraw() is abused as a hook into screen.c to remove a
potential softcursor.

Reference: /n/sources/patch/bad/softcurs-pixel-dropping
Date: Fri Mar 11 18:29:22 CET 2011
Signed-off-by: cinap_lenrek@gmx.de