patch: maybe/vga-vesa-modes
fix writing garbage cursor image in main by removing
the code. the kernel gc drivers will load the arrow
image when initialized so there is no need to touch it.

the vesa modeinfo dx field is the visual width of
the framebuffer image and is not always equal to the
width of the framebuffer memory image. account for this
by setting vga->virtx in the options phase when needed.

filter out modes in vbemodeinfo that are not supported
by plan9 devdraw (text, planar, non chain, yuv, stereo).

filter out modes that say they are not supported by the
hardware, dont support linear framebuffer or being non
graphics modes.

renamed the Vmode.name field to Vmode.size as thats what
it really is. fill in standard vesa timing information
in dbvesamode().

some other minjor fixes in vesa.c

Reference: /n/sources/patch/maybe/vga-vesa-modes
Date: Wed Mar 16 01:34:55 CET 2011
Signed-off-by: cinap_lenrek@gmx.de