patch: sorry/cifs-mtu-wrong
nasty bug which appeared to cause incorrect sequence numbers when accessing very
large directories or reading the contents of some of the virtual files in the root of the cifs dir.

This is caused by having a larger ovvered receive buffer than the packet size, which makes the
windows server send multiple reply fragments, sadly the layering of the protocols is broken
and these fragments have the same sequence number in the lower SMB layer (ugh).

The fix is to just to ensure the receive buffer is always smaller than MTU and thus you always
get a reply in a single fragment.

-Steve

Notes:
Fri Jul 29 17:18:40 EDT 2011 geoff
    just going to update to current source.

Reference: /n/sources/patch/sorry/cifs-mtu-wrong
Date: Thu Mar 17 16:54:26 CET 2011
Signed-off-by: steve@quintile.net
Reviewed-by: geoff