patch: applied/replica-c-resolve
When a conflict is being resolved using local copy replica/applylog -c should save the resolution to the database, but some code paths just skips over.

In the past, the workaround was to give all -c/-s to the pull script at once so log apply time is updated and the entries are never revisited. However, this stopped to work when plan9.log began to be recreated monthly, and the earlier-resolved conflicts were reported again and again.

This patch solves the problem for 'a' and 'c' verbs in the log.

Reference: /n/sources/patch/applied/replica-c-resolve
Date: Mon May 16 17:25:52 CES 2011
Signed-off-by: yarikos@gmail.com