patch: maybe/dist-network
Fix network installation option in installer.

This is very useful for installing Plan 9 from PXE in a local network.

The plan9.ini parameter "installurl" give an example on how to set the
Plan 9 CD image URL. By default, this option is set to the Bell Labs
server and can be used to download the last CD image from a floppy or
an older CD.

Reference: /n/sources/patch/maybe/dist-network
Date: Fri Nov 11 00:00:03 CET 2011
Signed-off-by: djc@9grid.fr