patch: maybe/dist-venti-bloom
Add Venti Bloom filter configuration to installer.

The size of the Bloom filter is calculated as the following:
  - the power of two, immediately inferior to 1/3 of 20/100
     of total memory.

Where:
 - 20/100 is the percentage of free memory allocated by Venti,
   as defined in /sys/src/9/boot/local.c,
 - 1/3 is the amount of Venti memory allocated to the Bloom filter.
And:
 - the result cannot be superior to the Bloom partition size,
 - the result cannot be inferior to 1 MB.

The method of calculation of the percentage of free memory allocated
by Venti is the same as the one defined in /sys/src/cmd/venti/venti.c.

We don't care about reducing the Bloom partition since disk space is
cheap these days and the unused space could always be used when more
memory is available.
On a common hard disk, the Bloom partition would be automatically
set to 512MB by disk/prep.

Probably you would like to make some adjustments. This is just
the general idea.
Reference: /n/sources/patch/maybe/dist-venti-bloom
Date: Thu Nov 17 15:10:55 CET 2011
Signed-off-by: djc@9grid.fr